added leaveRoom

+++ server2/routes/room.js
@@ -1,6 +1,7 @@
 import { Router } from "express"; // Express web server framework
 import url from "url";
 import Room from "../models/room";
+import User from "../models/user";
 import authenticate from "../middleware/auth";
 import _ from "lodash";
 import { v4 as uuid4 } from "uuid";
@@ -52,10 +53,34 @@ router.get("/delete", authenticate, async (req, res, next) => {
   const room = await Room.findOne({ id });
   if (!room) return next({ statusCode: 404, message: "room not found" });
 
-  Room.deleteOne({ id: room.id }, (err) => {
-    if (err) return next(err);
-    return res.status(200).json({ message: "room deleted succesfully" });
+  if (user.username === room.users[0].username)
+    Room.deleteOne({ id: room.id }, (err) => {
+      if (err) return next(err);
+      return res.status(200).json({ message: "room deleted succesfully" });
+    });
+  else return next({ message: "need admin priviledges", statusCode: 403 });
+});
+
+router.get("/leave", authenticate, async (req, res, next) => {
+  const url_parts = url.parse(req.url, true);
+  const { id } = url_parts.query;
+  const user = await User.findById(req.user._id);
+  if (!id) return next({ statusCode: 401, message: "no id given" });
+  const room = await Room.findOne({ id });
+  if (!room) return next({ statusCode: 404, message: "room not found" });
+
+  _.remove(room.users, {
+    username: user.username,
   });
+  room.markModified("users");
+  await room.save();
+
+  _.remove(user.rooms, (r) => {
+    return r.toString() === room._id.toString();
+  });
+  user.markModified("rooms");
+  await user.save();
+  return res.status(200).json({ message: "Leaved room succesfully" });
 });
 
 router.get("/create", authenticate, async (req, res, next) => {
@@ -79,7 +104,6 @@ router.post("/update", authenticate, async (req, res, next) => {
     return res.status(200).json(doc);
   });
 });
-
 
 router.get("/userRooms", authenticate, async (req, res, next) => {
   const user = req.user;
